connect through mongoose

+++ packages/server/setup/mongo-starter.js
@@ -1,5 +1,4 @@
 require("dotenv").config();
-const { spawn } = require("child_process");
 const { exec } = require("shelljs");
 
 class MongoStarter {
@@ -7,15 +6,17 @@ class MongoStarter {
     if (process.env.START_MONGO_DB === "true") {
       const { USE_AUTHENTICATION, MONGO_DB_PATH, MONGO_DB_PORT } = process.env;
       const command = `mongod ${
-        USE_AUTHENTICATION ? "--auth" : ""
+        USE_AUTHENTICATION === "true" ? "--auth" : ""
       } --port ${MONGO_DB_PORT} --dbpath ${MONGO_DB_PATH}`;
-      this.child = exec(command, { async: true });
+      this.child = exec(command, { async: true, silent: true });
     }
   }
 
   kill() {
-    exec("kill -2 `pgrep mongod`");
-    this.child.kill();
+    if (this.child) {
+      exec("kill -2 `pgrep mongod`");
+      this.child.kill();
+    }
   }
 }
